Refactoring of methods and functions for utilities modules
